requirements.txt
fastapi
uvicorn
python-multipart
PyPDF2
langchain
openai
faiss-cpu
pinecone-client
python-dotenv
langchain-community
